Re-enabled debug colours, partially fixed explosions
Ian, you don't seem to apply force on the Y-axis, just on the X. Any
reason?
Also, the glitchy explosions are caused, in part, by the direction of
iteration. The particles on the right clear out of the way, but on the
left, the sand is still in the way.